Fix dtype.

+++ sid/tests/test_contacts.py
@@ -198,7 +198,7 @@ def test_calculate_infections():
     )
 
     exp_infected = pd.Series([False] + [True] * 3 + [False] * 4)
-    exp_infection_counter = pd.Series([3] + [0] * 7)
+    exp_infection_counter = pd.Series([3] + [0] * 7).astype(np.int32)
     exp_immune = pd.Series([True] * 4 + [False] * 4)
     assert calc_infected.equals(exp_infected)
     assert calc_states["infection_counter"].equals(exp_infection_counter)
